% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg.R
\name{arg_index}
\alias{arg_index}
\title{Get Index Arguments from Call}
\usage{
arg_index(node)
}
\arguments{
\item{node}{A call with arguments.}
}
\value{
A list of index arguments.
}
\description{
This function gets the index arguments from calls that access slots, subset,
or replace an object.
}
