% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg.R
\name{arg_object}
\alias{arg_object}
\title{Get Object Argument from Call}
\usage{
arg_object(node)
}
\arguments{
\item{node}{A call with arguments.}
}
\value{
The object argument from the call.
}
\description{
This function gets the object argument from calls that access slots, subset,
or replace an object.
}
