% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg.R
\name{arg_value}
\alias{arg_value}
\title{Get Value Argument from Replacement Call}
\usage{
arg_value(node)
}
\arguments{
\item{node}{(Replacement) A call to a replacement function.}
}
\value{
The \code{value} argument from the call.
}
\description{
This function gets the value argument from calls that replace an object.
}
