% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_blocks.R
\name{as_blocks}
\alias{as_blocks}
\title{Convert Code Data Frame to Basic Blocks}
\usage{
as_blocks(code, new_blocks = FALSE)
}
\arguments{
\item{code}{(data.frame) The data frame to convert.}

\item{new_blocks}{(logical) Whether to create new Block objects. Beware that
these new blocks will become the new parents of the expressions in the
block.}
}
\description{
This function converts a code data frame to basic blocks.
}
