% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_blocks.R
\name{as.data.frame.BlockList}
\alias{as.data.frame.BlockList}
\alias{as.data.frame.Function}
\alias{as_data_frame}
\title{Convert Basic Blocks to a Code Data Frame}
\usage{
\method{as.data.frame}{BlockList}(x, ...)

\method{as.data.frame}{Function}(x, ...)

as_data_frame(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{(BlockList) The basic blocks to convert.}

\item{...}{Additional arguments to be passed to or from methods.}
}
\description{
This function converts basic blocks to a data frame where each row is one
``line'' of code.
}
