% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_igraph.R
\name{as_igraph}
\alias{as_igraph}
\title{Convert CFGraph to igraph}
\usage{
as_igraph(cfg)
}
\arguments{
\item{cfg}{(CFGraph) a graph to convert}
}
\description{
This function converts a CFGraph object to an igraph object, which is useful
for plotting the graph. The resulting igraph does not retain the code inside
the basic blocks, only the structure of the original CFGraph.
}

