% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_language.R
\name{to_r}
\alias{to_r}
\alias{as_language}
\title{Convert to R Language Objects}
\usage{
as_language(node, ...)
}
\arguments{
\item{node}{The intermediate representation to convert.}

\item{...}{Optional arguments to and from methods.}
}
\description{
This function converts rstatic intermediate representations back to R
language objects.
}
