% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_analysis.R
\name{backward_analysis}
\alias{backward_analysis}
\title{Compute Fixed Point Solution for Backward Analysis}
\usage{
backward_analysis(cfg, initial, killgen, confluence = union,
  update = dfa_standard_update, ..., full_analysis = FALSE,
  max_iter = 1000L)
}
\arguments{
\item{cfg}{(ControlFlowGraph) The CFG to compute the analysis for.}

\item{initial}{Initial guess for result set for each block.}

\item{killgen}{Kill and gen sets for each block.}

\item{update}{A function to compute the update set for a block.}

\item{...}{Additional arguments to the update function.}

\item{full_analysis}{(logical) If \code{TRUE}, the update sets are computed
in addition to the result sets.}

\item{max_iter}{(integer) Maximum number of iterations to run.}

\item{(list)}{A list that contains the result sets and (in a full analysis)
the update sets.}
}
\description{
This function solves for the maximal fixed point in a backward analysis.
}
