% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_analysis.R
\name{block_aggregate}
\alias{block_aggregate}
\title{Aggregate Data Flow Analysis Sets to Block-level}
\usage{
block_aggregate(gen, kill, forward = TRUE,
  update = dfa_standard_update)
}
\arguments{
\item{gen}{(logical matrix) Gen sets for each line.}

\item{kill}{(logical matrix) Kill sets for each line.}

\item{forward}{(logical) Should sets be aggregated forward (first to last
line) or backward (last to first line)?}

\item{update}{(function) A function to compute the update set for a block.}
}
\value{
A logical vector with \code{nrow(gen)} elements. The set computed by
aggregating over all of the line-level sets.
}
\description{
This function takes line-level sets from a data flow analysis and aggregates
them to block-level by sequential application of the update function.
}
