% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks_to_r.R
\name{blocks_to_r}
\alias{blocks_to_r}
\title{Generate R Code from Basic Blocks}
\usage{
blocks_to_r(node, ...)
}
\arguments{
\item{node}{The basic blocks to convert.}

\item{...}{Optional arguments to and from methods.}
}
\description{
This function converts basic blocks back to R code.
}
