% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlanguge.R
\name{body}
\alias{body}
\alias{body<-}
\title{Extract or Replace the Body of a Function Definition}
\usage{
body(fun)

body(fun, envir) <- value
}
\arguments{
\item{fun}{(Function) The function definition from which to extract or
replace the body.}

\item{envir}{Currently unused.}

\item{value}{The value to replace the body with.}
}
\description{
This function extracts or replaces the body of the given function
definition.
}
