% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_cfg.R
\name{build_cfg}
\alias{build_cfg}
\title{Build Basic Blocks from ASTNodes}
\usage{
build_cfg(node, builder)
}
\arguments{
\item{node}{(ASTNode) An ASTNode to build the graph from.}

\item{builder}{(CFGBuilder) The graph builder.}
}
\description{
This helper function does a depth-first traversal of an AST in order to
build basic blocks for a CFG.
}
\details{
Generally, this function should only be called from \code{to_cfg()}.
}
