% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfg_builder.R
\docType{data}
\name{CFGBuilder}
\alias{CFGBuilder}
\title{CFG Builder}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CFGBuilder
}
\description{
This class is used by \code{to_cfg()} to build a control flow graph.
}
\keyword{datasets}
