% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/child.R
\name{child}
\alias{child}
\alias{child<-}
\title{Extract or Replace Descendant of an RStatic Object}
\usage{
child(node, index)

child(node, index) <- value
}
\arguments{
\item{node}{(ASTNode) The node from which to extract or replace the
descendant.}

\item{index}{(integer) The index of the descendant to extract or replace.}

\item{value}{(ASTNode) The value to replace the descendant with.}
}
\value{
A single child node.
}
\description{
This function extracts or replaces a descendant of the given node.
}
\details{
If \code{index} is a vector, the \code{child} function is called recursively
so that descendants several levels below \code{node} are accessible.
}
\examples{
code = quote_ast(mean(1, 2, 3 + x))
x = child(code, c(2, 3, 2, 2))

child(code, c(2, 3, 2, 2)) = Integer$new(4)
}
\seealso{
\code{children} to extract or replace all children at once.
}
