% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/child.R
\name{child_fields}
\alias{child_fields}
\title{Fields Containing Children for an RStatic Object}
\usage{
child_fields(x)
}
\arguments{
\item{x}{(ASTNode) The node.}
}
\description{
Given an RStatic object, this function returns the names of fields which
contain children.
}
\details{
This function is the basis for most of the functions that extract or replace
children from RStatic objects.
}
\seealso{
\code{ls} to get the names of all fields.
}
