% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/child.R
\name{children}
\alias{children}
\alias{children<-}
\title{Extract or Replace All Children of an RStatic Object}
\usage{
children(node)

children(node) <- value
}
\arguments{
\item{node}{(ASTNode) The node from which to extract or replace the
children.}

\item{value}{(list) A named list of objects to replace the children with.
Names should match field names of \code{node}.}
}
\value{
A list of the child nodes.
}
\description{
This function extracts or replaces all children of the given node.
}
\examples{
code = quote_ast(x <- 3)
children(code)

children(code) = list(read = Integer$new(42), write = Symbol$new("y"))
}
\seealso{
\code{child} to extract or replace a single, specific descendant
slightly more efficiently.
}
