% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{class_error_string}
\alias{class_error_string}
\title{Create Error String from Object Class}
\usage{
class_error_string(object, msg = "Invalid class '\%s'.",
  all_classes = FALSE)
}
\description{
This function creates an error string that includes the class of an object.
}
