% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_namespace.R
\name{collapse_namespace}
\alias{collapse_namespace}
\title{Collapse Namespace}
\usage{
collapse_namespace(node, ...)
}
\arguments{
\item{node}{(ASTNode) The node to collapse.}

\item{...}{Unused arguments.}
}
\description{
If \code{node} is a Namespace (a call to \code{::} or \code{:::}), this
function returns the Symbol the Namespace affects and stores associated
namespace information in the \code{namespace} field of that Symbol. If
\code{node} is not a Namespace, this function returns \code{node} unchanged.
}
\details{
This function is intended as a replace function for \code{replace_nodes()}.
}
\examples{
code = quote_ast(rstatic::collapse_one_namespace)
replace_nodes(code, collapse_namespace)
}
