% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_namespaces.R
\name{collapse_namespaces}
\alias{collapse_namespaces}
\title{Collapse Namespaces}
\usage{
collapse_namespaces(node, in_place = FALSE)
}
\arguments{
\item{node}{(ASTNode) The AST.}

\item{in_place}{(logical) Whether the AST should be copied before
transformation.}
}
\description{
This function collapses namespaces in the AST, so that they are stored in
the namespace field of each Symbol rather than as a call to \code{::} or
\code{:::}.
}
