% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_namespaces.R
\name{collapse_node_namespaces}
\alias{collapse_node_namespaces}
\title{Collapse A Namespace Node}
\usage{
collapse_node_namespaces(node, ...)
}
\arguments{
\item{node}{(ASTNode) The ASTNode to collapse.}
}
\description{
This function will collapse a Namespace node onto the Symbol in its second
argument. Typically this will be used with \code{node_apply()} or through
the frontend \code{collapse_namespaces()}.
}
