% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_namespace.R
\name{collapseNamespaces}
\alias{collapseNamespaces}
\title{Collapse Namespaces}
\usage{
collapseNamespaces(node, inPlace = FALSE)
}
\arguments{
\item{node}{(ASTNode) The AST.}

\item{inPlace}{(logical) Whether the AST should be copied before
transformation.}
}
\description{
This function collapses namespaces in the AST, so that they are stored in
the namespace field of each Symbol rather than as a call to \code{::} or
\code{:::}.
}
