% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_namespace.R
\name{collapseNodeNamespaces}
\alias{collapseNodeNamespaces}
\title{Collapse A Namespace Node}
\usage{
collapseNodeNamespaces(node, ...)
}
\arguments{
\item{node}{(ASTNode) The ASTNode to collapse.}
}
\description{
This function will collapse a Namespace node onto the Symbol in its second
argument. Typically this will be used with \code{nodeApply()} or through the
frontend \code{collapseNamespaces()}.
}
