% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfg_analysis.R
\name{collect_crossblock_uses}
\alias{collect_crossblock_uses}
\title{Collect Cross-block Uses in a CFG}
\usage{
collect_crossblock_uses(cfg)
}
\arguments{
\item{cfg}{(CFGraph) A control flow graph.}
}
\description{
This function collects the (unique) names of all variables that are used in
a different block than the block where they were defined.
}
