% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast_analysis.R
\name{collectReads}
\alias{collectReads}
\title{Collect All Reads in an AST}
\usage{
collectReads(node)
}
\arguments{
\item{node}{(ASTNode) An abstract syntax tree.}
}
\description{
This function collects the (unique) names of all variables that are read in
an AST. This function doesn't consider calling a function a read.
}
\details{
Beware that there are many ASTNode subclasses that this function doesn't
attempt to handle; the function raises an error for unsupported classes.
}
