% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{:=}
\alias{:=}
\title{Unpack Elements}
\usage{
":="(x, y)
}
\description{
This operator unpacks the elements of a vector or list into variables. Since
many other packages provide their own definition of the \code{:=} operator, this
function is not exported and only meant for internal use.
}
