% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_analysis.R
\name{compute_gen_kill}
\alias{compute_gen_kill}
\title{Compute Gen and Kill Sets for Data Flow Analysis}
\usage{
compute_gen_kill(cfg, fun, universe, ...)
}
\arguments{
\item{cfg}{(BlockList) A control flow graph on which to compute gen and kill
sets.}

\item{fun}{(function) A function which accepts an \code{ASTNode} and returns a
list with elements named \code{gen} and \code{kill}. Each element should be a logical
vector with the same length as \code{universe}.}

\item{universe}{(named list) The universal set, which contains all items
used in the analysis.}

\item{...}{Additional arguments to \code{fun}.}
}
\description{
This function helps with computing the gen and kill sets for a data flow
analysis.
}
\seealso{
\code{\link[=forward_analysis]{forward_analysis()}}
}
