% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy.R
\name{copy}
\alias{copy}
\title{Copy an RStatic Object}
\usage{
copy(x, ...)
}
\arguments{
\item{x}{The object to copy.}

\item{...}{Additional arguments to methods.}

\item{skip_set_parent}{(character) Names of fields which should never have
\code{set_parent} called on them.}
}
\description{
This function copies the given RStatic object, while ensuring that
parent-child relationships are preserved for the copied object.
}
\details{
If \code{x} is any other R6 object, \code{x} is deep-cloned. If \code{x} is
not an R6 object, no action is taken and \code{x} is returned.

Since RStatic objects are references, assignment does not make a copy. This
function can be used to explicitly copy an RStatic object.
}
\examples{
x = quote_ast(x <- "Hi DTL!")
y = x
z = copy(x)

x$read = Numeric$new(141)

# Changing 'x' changed 'y' (a reference), but not 'z' (a copy).
}
\seealso{
\code{set_parent}, which is used by this function.
}
