% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counter.R
\docType{data}
\name{Counter}
\alias{Counter}
\title{Counter for Strings}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Counter
}
\description{
This data structure keeps a running total of the number of times it has seen
each unique string. This is an R6 class, so the object can be modified in
place and assignment has reference semantics.
}
\keyword{datasets}
