% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_analysis.R
\name{dfa_gen_kill}
\alias{dfa_gen_kill}
\title{Compute Gen and Kill Sets for Data Flow Analysis}
\usage{
dfa_gen_kill(cfg, fun, universe, forward = TRUE, ...)
}
\arguments{
\item{cfg}{(BlockList) A control flow graph on which to compute gen and kill
sets.}

\item{fun}{(function) A function which accepts an \code{ASTNode} and returns a
list with elements named \code{gen} and \code{kill}. Each element should be a logical
vector with the same length as \code{universe}.}

\item{universe}{(named list) The universal set, which contains all items
used in the analysis.}

\item{...}{Additional arguments to \code{fun}.}
}
\value{
A two-element list. The first element, "gen", is a logical matrix
that represents the gen set for each block. The second element, "kill", is
a logical matrix that represents the kill set for each block. For both
matrices, rows correspond to set items and columns correspond to blocks.
}
\description{
This function helps with computing the gen and kill sets for a data flow
analysis.
}
\seealso{
\code{\link[=dfa_solve]{dfa_solve()}}
}
