% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_analysis.R
\name{dfa_initialize}
\alias{dfa_initialize}
\title{Initial Sets for Data Flow Analysis}
\usage{
dfa_initialize(cfg, universe, default = FALSE, complement = c("none",
  "entry", "exit"))
}
\arguments{
\item{cfg}{(BlockList) A control flow graph for which to initialize the
sets.}

\item{universe}{(named list) The universal set, which contains all items
used in the analysis.}

\item{default}{(logical) If \code{TRUE}, sets are initialized to the universal
set. Otherwise, sets are initialized to the empty set.}

\item{complement}{(character) A block which should be initialized to the
complement of the default set.}
}
\value{
A logical matrix that represents the solution set of each block.
Rows correspond to set items and columns correspond to blocks.
}
\description{
This function generates initial sets for a data flow analysis.
}
\seealso{
\code{\link[=dfa_solve]{dfa_solve()}}
}
