% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_analysis.R
\name{dfa_propagate}
\alias{dfa_propagate}
\title{Propagate Data Flow Analysis Result Sets within a Block}
\usage{
dfa_propagate(initial, gk, forward = TRUE,
  update = dfa_standard_update, ...)
}
\arguments{
\item{initial}{(logical matrix) Initial set to propagate. For forward
propagation, this is the entry set for the block. For backward propagation,
this is the exit set for the block.}

\item{forward}{(logical) Should sets be aggregated forward (first to last
line) or backward (last to first line)?}

\item{update}{(function) A function to compute the update set for a block.}

\item{gen}{(logical matrix) Gen sets for each line.}

\item{kill}{(logical matrix) Kill sets for each line.}
}
\value{
A logical matrix with \code{nrow(gen)} rows and \code{ncol(gen) + 1} columns.
Rows correspond to set items. Columns represent entry sets for lines; the
last column is the exit set for the last line (equivalently, for the block).
}
\description{
This function propagates result sets from a data flow analysis at line-level
by sequential application of the update function.
}
\seealso{
\code{\link[=dfa_aggregate]{dfa_aggregate()}}, \code{\link[=dfa_disaggregate]{dfa_disaggregate()}}, \code{\link[=dfa_standard_update]{dfa_standard_update()}}
}
