% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_analysis.R
\name{dfa_standard_update}
\alias{dfa_standard_update}
\title{Standard Update for Data Flow Analyses}
\usage{
dfa_standard_update(result, gen, kill, ...)
}
\arguments{
\item{result}{The result set for the current block.}

\item{gen}{The gen set for the current block.}

\item{kill}{The kill set for the current block.}

\item{...}{Additional optional arguments to the update function.}
}
\value{
The update set for the next block.
}
\description{
This function computes the standard update set for data flow analyses. The
standard update set is the result set less the kill set, all unioned with
the gen set.
}
\seealso{
\code{\link[=dfa_solve]{dfa_solve()}}, \code{\link[=dfa_aggregate]{dfa_aggregate()}}
}
