% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominators.R
\name{dom_tree}
\alias{dom_tree}
\title{Compute Dominator Tree}
\usage{
dom_tree(cfg, entry_idx = cfg$entry_index())
}
\arguments{
\item{cfg}{(ControlFlowGraph) A control-flow graph.}
}
\value{
The dominator tree as a vector of immediate dominators. In other
words, if element \eqn{j} is \eqn{i}, then the immediate dominator of block
\eqn{j} is block \eqn{i}. The dominator set for block \eqn{j} consists of
all blocks on the path from block \eqn{j} to the root of the dominator tree.
}
\description{
This function computes the dominator tree for a control-flow graph.
}
\details{
A block \eqn{b_i} dominates another block \eqn{b_j} when all program paths
to \eqn{b_j} must pass through \eqn{b_i}. The dominance property is
reflexive: every block dominates itself. This function iteratively computes
the immediate dominator for each block until a fixed point is reached.
}
\references{
Cooper, K. D., Harvey, T. J., and Kennedy, K. (2001) A simple, fast
dominance algorithm. Software Practice & Experience 4, 1-10.

Cooper, K. D. and Torczon, L. (2012) Engineering a Compiler. Elsevier.
}
