% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominators.R
\name{domFrontier}
\alias{domFrontier}
\title{Compute Dominance Frontiers}
\usage{
domFrontier(cfg, domTree)
}
\arguments{
\item{cfg}{(ControlFlowGraph) A control-flow graph.}

\item{domTree}{(integer) The dominator tree for the control-flow graph.}
}
\value{
The dominance frontiers as a list of integer vectors. Each element
of the list is the dominance frontier for the corresponding block in the
control-flow graph.
}
\description{
This function computes the dominance frontiers for a control-flow graph.
}
\details{
The dominance frontier for a block \eqn{b} is the set of all nodes \eqn{y}
such that \eqn{b} dominates are predecessor of \eqn{y} but does not strictly
dominate \eqn{y}. In other words, the dominance frontier for \eqn{b} is the
set of blocks immediately beyond the blocks dominated by \eqn{b}, where
control-flow merges from a separate part of the program.
}
\references{
Cooper, K. D., Harvey T. J., and Kennedy, K. (2001) A simple, fast dominance
algorithm. Software Practice & Experience 4, 1-10.

Cooper, K. D. and Torczon, L. (2012) Engineering a Compiler. Elsevier.
}
