% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph.R
\name{dominator_tree}
\alias{dominator_tree}
\title{Dominator Tree of a Graph}
\usage{
dominator_tree(g, root = 2)
}
\arguments{
\item{g}{(igraph) The graph.}

\item{root}{(integer) Index of the root node of the graph.}
}
\value{
The dominator tree as an igraph graph.
}
\description{
This function computes the dominator tree for a graph.
}
\details{
This is a utility function to make working with igraph easier.

A block \eqn{b_i} dominates another block \eqn{b_j} when all paths to
\eqn{b_j} must pass through \eqn{b_i}. The closest dominator of a block
(excluding the block itself) is called its immediate dominator. In the
dominator tree, each block is a child of its immediate dominator.
}
