% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominators.R
\name{dominators}
\alias{dominators}
\title{Compute Dominator Sets}
\usage{
dominators(cfg)
}
\arguments{
\item{cfg}{(CFGraph) a control-flow graph}
}
\description{
This function computes the dominator sets for each block in a control-flow
graph.
}
\details{
A block \eqn{b_i} dominates another block \eqn{b_j} when all program paths
to \eqn{b_j} must pass through \eqn{b_i}. The dominance property is
reflexive: every block dominates itself. This function iteratively computes
the set of dominators for each block until a fixed point is reached. In most
cases, only 2-3 iterations are necessary.
}

