% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominators.R
\name{domTreePreorder}
\alias{domTreePreorder}
\title{Preorder Traversal of a Tree}
\usage{
domTreePreorder(domTree, from = 1L)
}
\arguments{
\item{domTree}{(integer) The tree on which to compute a traversal, in vector
form. The value at each index should be the index of that node's parent.}

\item{from}{(integer) A node index where traversal should start.}
}
\value{
(integer) The node indexes in the order they'd be visited.
}
\description{
Given a tree in vector form, this function finds the order nodes are visited
by a preorder traversal.
}
