% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_blocklist.R
\name{entry_block}
\alias{entry_block}
\title{Get Entry Block}
\usage{
entry_block(node)
}
\arguments{
\item{node}{(Function | BlockList) An object which contains Blocks.}
}
\value{
The entry block.
}
\description{
This function gets the entry block from a Function or BlockList.
}
