% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equal.R
\name{equal}
\alias{equal}
\alias{==.ASTNode}
\alias{!=.ASTNode}
\title{Test Equality of RStatic Objects}
\usage{
equal(x, y)

\method{==}{ASTNode}(x, y)

\method{!=}{ASTNode}(x, y)
}
\arguments{
\item{x, y}{(ASTNode) The objects to compare.}
}
\value{
A logical scalar indicating the result of the comparison.
}
\description{
These functions and operators allow comparison of RStatic objects.
}
\details{
These functions traverse the given RStatic objects to check that their
children are also equal. Except for Symbol names and Literal values, fields
that do not contain children are ignored.
}
\examples{
x = code_objects_ast(a <- 3)
y = code_objects_ast(a <- 3)
z = code_objects_ast(a <- 4)

x == y

y == z
}
