% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_blocklist.R
\name{exit_block}
\alias{exit_block}
\title{Get Exit Block}
\usage{
exit_block(node)
}
\arguments{
\item{node}{(Function | BlockList) An object which contains Blocks.}
}
\value{
The exit block.
}
\description{
This function gets the exit_block from a Function or BlockList.
}
