% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_call.R
\name{find_function}
\alias{find_function}
\alias{find_function.call}
\alias{find_function.Call}
\title{Find Function Definition for Call}
\usage{
find_function(node, envir, top)

\method{find_function}{call}(node, envir = parent.frame(), top = TRUE)

\method{find_function}{Call}(node, envir = parent.frame(), top = TRUE)
}
\arguments{
\item{node}{(call|Call) The unevaluated call.}

\item{envir}{(environment) An environment to look up the definition.}

\item{top}{(logical) Is this the top-level call to \code{find_function()}? This
parameter should not be set by the user.}
}
\value{
A function.
}
\description{
This function returns the definition of the function to be called by an
unevaluated call.
}
\details{
This function raises an error if it is unable to find the definition.
}
