% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_functions.R
\name{find_functions}
\alias{find_functions}
\title{Find Function Definitions in a Code Object}
\usage{
find_functions(node, initial = list(), ..., recursive = FALSE)
}
\arguments{
\item{node}{(ASTNode) The code object to search.}

\item{initial}{(list) An initial list of functions.}

\item{...}{Additional arguments to and from methods.}

\item{recursive}{(logical) Should function definitions in \code{node} also
be searched?}
}
\description{
This function takes a code object and returns a list of all function
definitions.
}
