% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_nodes.R
\name{find_node_indices}
\alias{find_node_indices}
\title{Find Node Indices in an RStatic Object}
\usage{
find_node_indices(node, test, ..., recursive = TRUE, initial = list(),
  index = integer(0))
}
\arguments{
\item{node}{(ASTNode) The RStatic object to search.}

\item{test}{(function) The test function, which should accept an rstatic
object to check as its first argument and return a scalar logical.}

\item{...}{Additional arguments to the test function.}

\item{recursive}{(logical) Search recursively in matching nodes?}

\item{initial}{(list) Initial list of indices for matching nodes.}

\item{index}{(integer) The index of \code{node}.}
}
\value{
A list of indices for matching nodes.
}
\description{
This function traverses the given RStatic object and returns a list of
indices for all descendants that satisfy the given test function.
}
\details{
All indices are relative to \code{node}, unless \code{index} is supplied.

Note that \code{node} itself is not tested.
}
\examples{
ast = quote_ast(
  function(x, y) {
    x = x + y
    function() x
  })

find_node_indices(ast, is, "Function")

find_node_indices(ast, is_symbol, "x")
}
\seealso{
\code{\link[=find_nodes]{find_nodes()}}
}
