% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_nodes.R
\name{find_nodes}
\alias{find_nodes}
\title{Find Nodes in an RStatic Object}
\usage{
find_nodes(node, test, ..., recursive = TRUE, initial = list())
}
\arguments{
\item{node}{(ASTNode) The RStatic object to search.}

\item{test}{(function) The test function, which should accept an rstatic
object to check as its first argument and return a scalar logical.}

\item{...}{Additional arguments to the test function.}

\item{recursive}{(logical) Search recursively in matching nodes?}

\item{initial}{(list) Initial list of matching nodes.}
}
\value{
A list of matching nodes.
}
\description{
This function traverses the given RStatic object and returns a list of all
descendants that satisfy the given test function.
}
\examples{
ast = quote_ast(
  function(x, y) {
    x = x + y
    function() x
  })

find_nodes(ast, is, "Function")

find_nodes(ast, is_symbol, "x")
}
\seealso{
\code{\link[=find_node_indices]{find_node_indices()}}, \code{\link[=replace_nodes]{replace_nodes()}}
}
