% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlanguge.R
\name{formals}
\alias{formals}
\alias{formals<-}
\title{Extract or Replace the Parameters of a Function Definition}
\usage{
formals(fun)

formals(fun, envir) <- value
}
\arguments{
\item{fun}{(Function) The function definition from which to extract or
replace the parameters.}

\item{envir}{Currently unused.}

\item{value}{The value to replace the parameters with.}
}
\description{
This function extracts or replaces the parameters (formal arguments) of the
given function definition.
}
