% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_analysis.R
\name{forward_analysis}
\alias{forward_analysis}
\title{Compute Fixed Point Solution for Forward Analysis}
\usage{
forward_analysis(cfg, initial, gen, kill, confluence = `|`,
  update = dfa_standard_update, ..., full_analysis = FALSE,
  max_iter = 1000L)
}
\arguments{
\item{cfg}{(BlockList) A control flow graph with which to compute a
solution.}

\item{initial}{Initial guess of solution for each block.}

\item{gen}{Gen sets for each block.}

\item{kill}{Kill sets for each block.}

\item{confluence}{(function) A function to combine old set and update set
for a block.}

\item{update}{(function) A function to compute the update set for a block.}

\item{...}{Additional arguments to \code{update}.}

\item{full_analysis}{(logical) If \code{TRUE}, the update sets are computed in
addition to the result sets.}

\item{max_iter}{(integer) Maximum number of iterations to run.}
}
\value{
A two-element list. The first element, "entry", is a list of
solution sets at the entry from each block. The second element, "exit", is
\code{NULL}, or for a full analysis, a list of solution sets at the exit to each
block.
}
\description{
This function solves for the maximal fixed point in a forward analysis.
}
\seealso{
\code{\link[=backward_analysis]{backward_analysis()}}
}
