% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_index.R
\name{get_index}
\alias{get_index}
\title{Get Index Arguments From Subset Call}
\usage{
get_index(node)
}
\arguments{
\item{node}{(Subset) A call to a subset function.}
}
\value{
A list of index arguments.
}
\description{
This function gets the index arguments from a call to a subset function.
}
