% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_symbols.R
\name{get_symbols}
\alias{get_symbols}
\title{Get Sets of Defined and Used Symbols}
\usage{
get_symbols(node, initial = SymbolSets(), ...)
}
\arguments{
\item{node}{(ASTNode) The expression to analyze.}

\item{initial}{(SymbolSets) The initial sets of defined/used expressions.}

\item{...}{Further arguments to methods.}
}
\value{
A \code{SymbolSets} object, with elements \code{def} and \code{use} that list
references to expressions that define or use symbols, respectively.
}
\description{
This function returns the set of symbols defined by an expression and the
set of symbols used by an expression.
}
\seealso{
\code{\link[CodeDepends:getInputs]{CodeDepends::getInputs()}}
}
