% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast_node_methods.R
\name{has_phi}
\alias{has_phi}
\title{Check Presence of Phi Function in Block}
\usage{
has_phi(block, x)
}
\arguments{
\item{block}{(Block) The block search for phi-functions.}

\item{x}{(character) The basename to match against phi-functions in the
block.}
}
\description{
This function checks whether the given basic block contains a phi-function
for the given basename. This function is used internally to avoid duplicating
phi-functions during SSA construction, but may also be useful for determining
whether the value of the variable arrives at the block from more than one
execution path.
}
