% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_return.R
\name{insert_return}
\alias{insert_return}
\title{Insert Optional Return Expressions}
\usage{
insert_return(node, ..., recursive = TRUE)
}
\arguments{
\item{node}{(Function) The function to transform.}

\item{...}{Optional arguments to and from methods.}

\item{recursive}{(logical) Should functions defined in \code{node} also be
transformed?}
}
\description{
This function transforms a function so that every expression that can be
returned is explicitly marked by a \code{return()} expression.
}
\details{
This is an in-place transformation! This function may not behave as expected
if called on a non-function.
}
