% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast_node_methods.R
\name{is_control_flow}
\alias{is_control_flow}
\title{Test If Node Is Control Flow}
\usage{
is_control_flow(node)
}
\arguments{
\item{node}{(ASTNode) The node to test.}
}
\description{
This function tests if a node is a control flow node.
}
