% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is_r6_method}
\alias{is_r6_method}
\title{Check Whether Named Member of R6 Object is Method}
\usage{
is_r6_method(name, obj)
}
\description{
This function checks whether the named members of an R6 object are methods
(as opposed to fields or active bindings).
}
