% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_symbol}
\alias{is_symbol}
\title{Is Object a Symbol?}
\usage{
is_symbol(x, name = NA_character_, ssa = TRUE)
}
\arguments{
\item{x}{The object to check.}

\item{name}{(character) A name to compare against, or \code{NA} to skip checking
the name (the default).}
}
\value{
A logical value: is \code{x} a Symbol, and does it match the \code{name}
argument?
}
\description{
This function checks whether the \code{x} argument is an rstatic Symbol. The
optional second argument can be used to check for a specific name.
}
\examples{
x = quote_ast(hi)
is_symbol(x) # TRUE
is_symbol(x, "hi") # TRUE
is_symbol(x, "hello") # FALSE
}
