% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{list_dots_safely}
\alias{list_dots_safely}
\title{Collect Dots, Preserving Missing Arguments}
\usage{
list_dots_safely(...)
}
\description{
This function collects the \code{...} argument into a list. Missing arguments are
preserved as S3 objects of class \code{missing}.
}
\seealso{
\code{\link[rlang:dots_list]{rlang::dots_list()}}
}
