% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_ast.R
\name{list_formals}
\alias{list_formals}
\title{List Formal Arguments}
\usage{
list_formals(expr)
}
\description{
This function returns the list of formal arguments, also called parameters,
for a function object.
}
\details{
For primitives that have no named parameters, this function returns an empty
list.
}
