% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/live_variables.R
\name{live_variables}
\alias{live_variables}
\title{Live Variables Analysis}
\usage{
live_variables(cfg, ...)
}
\arguments{
\item{cfg}{(ControlFlowGraph) The control flow graph to compute the live
variables analysis for.}

\item{...}{Additional arguments to \code{backward_analysis()}.}
}
\description{
This function computes the live variables at the exit of each basic block. A
variable is live if it has already been defined and it will be used in some
subsequent block.
}
