% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/live_variables.R
\name{live_variables_killgen}
\alias{live_variables_killgen}
\title{Compute Kill and Gen Sets for Live Variables Analysis}
\usage{
live_variables_killgen(node, initial)
}
\arguments{
\item{node}{(ASTNode) The basic block to compute}

\item{initial}{Initial}

\item{result}{}
}
\description{
Given a basic block, this function computes the kill and gen sets for live
variables analysis.
}
