% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{match_object}
\alias{match_object}
\title{Match Object(s) in List}
\usage{
match_object(x, table, nomatch = NA_integer_)
}
\arguments{
\item{x}{list or object: The object(s) to be matched.}

\item{table}{list: The objects to be matched against.}

\item{nomatch}{The value to be returned in the case when no match is found.}
}
\value{
An integer vector giving the position in \code{table} of the first match
if there is a match, otherwise \code{nomatch}.
}
\description{
This function returns a vector of the positions of the first matches of its
first argument in its second.
}
\seealso{
\code{\link[base:match]{base::match()}}
}
