% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssa_builder.R
\docType{data}
\name{NameGenerator}
\alias{NameGenerator}
\title{Generate Unique Variable Names}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
NameGenerator
}
\description{
Generate Unique Variable Names
}
\keyword{datasets}
