% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_symbols.R
\name{names_used}
\alias{names_used}
\title{Symbols Used by an Expression}
\usage{
names_used(x, ...)
}
\arguments{
\item{x}{A language object, \code{ASTNode} object, or \code{SymbolSets} object.}

\item{...}{Further arguments to methods.}
}
\value{
A character vector.
}
\description{
This function returns a character vector of the symbols used by an
expression.
}
\details{
If you need both the symbols defined and the symbols used by an expression,
it is more efficient to first call \code{get_symbols()} on the expression and
then call \code{names_defined()} and/or \code{names_used()} on the result.
}
\seealso{
\code{\link[=names_defined]{names_defined()}}, \code{\link[=get_symbols]{get_symbols()}}, \code{\link[CodeDepends:getInputs]{CodeDepends::getInputs()}}
}
