% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/child.R
\name{nchildren}
\alias{nchildren}
\title{The Number of Children of an RStatic Object}
\usage{
nchildren(x)
}
\arguments{
\item{x}{(ASTNode) The node.}
}
\description{
This function returns the number of children of the given node.
}
