% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.CFGraph}
\alias{plot.CFGraph}
\title{Plot Method for Control-flow Graphs}
\usage{
\method{plot}{CFGraph}(x, ...)
}
\arguments{
\item{x}{(CFGraph) A control-flow graph.}

\item{...}{Additional arguments to \code{plot.igraph}.}
}
\description{
This method plots a control flow graph using the igraph package.
}
