% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_graph.R
\name{plot.FlowGraph}
\alias{plot.FlowGraph}
\title{Plot Method for Flow Graphs}
\usage{
\method{plot}{FlowGraph}(x, ...)
}
\arguments{
\item{x}{(FlowGraph) A flow graph.}

\item{...}{Additional arguments to \code{plot.igraph}.}
}
\description{
This method plots a flow graph.
}
