% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfg_analysis.R
\name{postorder}
\alias{postorder}
\title{Postordering of a Graph}
\usage{
postorder(g, from = g$entry)
}
\arguments{
\item{g}{(FlowGraph) The graph on which to compute a traversal.}

\item{from}{(integer) A node index where the traversal should start.}
}
\value{
(integer) The node indexes in the order they'd be visited.
}
\description{
Given a graph, this function finds a postordering of the nodes.
}
\details{
A postordering gives nodes in the order they were last visited by a
depth-first search (as opposed to first visited). That is, the deepest nodes
appear first in the postordering. A postordering is not necessarily unique.

For an acyclic graph, a reversed postordering is equivalent to a topological
sort.
}
