% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph.R
\name{predecessors}
\alias{predecessors}
\title{Vertex Predecessors}
\usage{
predecessors(g, id)
}
\arguments{
\item{g}{(igraph) The graph.}

\item{id}{An index into the graph, or a Block.}
}
\description{
This function gets the predecessors of a vertex in a graph.
}
\details{
This is a utility function to make working with igraph easier.
}
