% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constant_propagation.R
\name{propagate_constants}
\alias{propagate_constants}
\title{Sparse Conditional Constant Propagation}
\usage{
propagate_constants(cfg)
}
\arguments{
\item{cfg}{(ControlFlowGraph) The graph on which to apply the algorithm.}
}
\value{
(list) A named list, with values for constant symbols and NONCONST
objects for nonconstant symbols.
}
\description{
This function applies sparse conditiona constant propagation to the
specified control flow graph.
}
\references{
Wegman, M. N., & Zadeck, F. K. (1991). Constant propagation with conditional
branches. ACM Transactions on Programming Languages and Systems (TOPLAS),
13(2), 181-210.
}
