% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_blocks.R
\name{quote_blocks}
\alias{quote_blocks}
\title{Convert an R Expression to Basic Blocks}
\usage{
quote_blocks(expr, ...)
}
\arguments{
\item{expr}{An unquoted R expression.}

\item{...}{Additional arguments to \code{to_blocks()}.}
}
\description{
This function converts an unquoted R expression to basic blocks.
}
