% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_cfg.R
\name{quote_cfg}
\alias{quote_cfg}
\title{Build Control Flow Graph from R Expression}
\usage{
quote_cfg(expr, ...)
}
\arguments{
\item{expr}{An unquoted R expression.}

\item{...}{Additional arguments to \code{to_cfg()}.}
}
\description{
This function builds the control flow graph (CFG) for an unquoted R
expression.
}
