% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{replace_in_parent}
\alias{replace_in_parent}
\title{Replace a Node in its Parent Node}
\usage{
replace_in_parent(node, replacement, in_place = TRUE)
}
\arguments{
\item{node}{(ASTNode) The node to replace.}

\item{replacement}{(ASTNode) The replacement node.}

\item{in_place}{(logical) If \code{FALSE}, parent of \code{node} is copied before
replacement.}
}
\value{
The parent node, after replacement.
}
\description{
This function replaces an ASTNode object in its parent expression with
another ASTNode object.
}
\details{
With the default arguments, this is an in-place transformation and has side
effects!

Only the first node \code{identical()} to \code{node} is replaced.
}
\examples{
node = quote_ast(x <- a + b)
ab = node$read
replace_in_parent(node, Symbol$new("t"))
}
