% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_nodes.R
\name{replace_nodes}
\alias{replace_nodes}
\title{Replace Nodes in an RStatic Object}
\usage{
replace_nodes(node, replace, ..., in_place = FALSE)
}
\arguments{
\item{node}{(ASTNode) The RStatic object to traverse.}

\item{replace}{(function) The function to call on each node.}

\item{...}{Additional arguments to the replace function.}

\item{in_place}{(logical) Copy \code{node} before replacement?}
}
\description{
This function traverses the given RStatic object, calls the given replace
function on each node, and replaces each node with the result of the call.
}
\details{
Replacement happens from the bottom up, so \code{node} is replaced last.
}
\examples{
rename_symbols = function(node, name, newname)
{ # Rename symbols to something else.
  if (is(node, "Symbol") && node$value == name)
    node$value = newname

  node
}

ast = quote_ast(ans <- sum(x, y / x, z))

replace_nodes(ast, rename_symbols, "x", "newx")
}
\seealso{
\code{\link[=find_nodes]{find_nodes()}}
}
