% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{row_ors}
\alias{row_ors}
\alias{row_ands}
\title{Form Row Disjunctions and Conjunctions}
\usage{
row_ors(x)

row_ands(x)
}
\description{
Form row disjunctions (logical \code{or}) and conjunctions (logical \code{and}) for
logical arrays.
}
\examples{
m = matrix(TRUE, 3, 2)
m[, 1L] = FALSE
row_ors(m)
row_ands(m)
}
