% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy.R
\name{set_parent}
\alias{set_parent}
\title{Set Parent of an RStatic Object}
\usage{
set_parent(node, parent)
}
\arguments{
\item{node}{The object on which to set the parent.}

\item{parent}{The new parent for \code{node}.}
}
\description{
This function sets the parent field for the given RStatic object or list of
RStatic objects.
}
\details{
If \code{node} is an RStatic object, then \code{set_parent(node, parent)} is
equivalent to \code{node$parent = parent}.

If \code{node} is a list, \code{set_parent} is called on each element.
}
