% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_blocks.R
\name{split_blocks}
\alias{split_blocks}
\title{Split Blocks at Control Flow}
\usage{
split_blocks(node)
}
\arguments{
\item{node}{(ASTNode) The root of the AST to be split.}
}
\description{
This function splits every block (Brace) in an AST after each control flow
expression. In the returned AST, a block can only have a control flow
expression as its final element.
}
