% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssa.R
\name{ssa}
\alias{ssa}
\title{Convert CFG Code to Static Single-Assignment Form}
\usage{
ssa(cfg)
}
\arguments{
\item{cfg}{(CFGraph) A control flow graph.}
}
\value{
The control flow graph as a CFGraph object, with the code in each
block converted to SSA form.
}
\description{
This function converts code in a control flow graph (CFG) to static
single-assignment form.
}

