% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_ssa.R
\name{ssa_insert_phis}
\alias{ssa_insert_phis}
\title{Insert SSA Phi-Functions}
\usage{
ssa_insert_phis(node, defs_by_block, dominators, live)
}
\description{
This function inserts SSA phi-functions into Blocks in code.
}
\details{
Generally, this function should only be called from \code{to_ssa()}.
}
