% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_ssa.R
\name{ssa_rename}
\alias{ssa_rename}
\title{Rename CFG Variables with SSA Names}
\usage{
ssa_rename(block, cfg, dom_t, builder, paramNames = character())
}
\arguments{
\item{block}{(integer) Index of a basic block in the CFG.}

\item{cfg}{(CFGraph) A control-flow graph.}

\item{dom_t}{(integer) The dominator tree for the CFG.}

\item{builder}{(SSABuilder) A stateful object used by the renaming}

\item{paramNames}{(character) A vector identifying the names of the parameters.
algorithm.}
}
\description{
This function renames variables in the basic blocks of a CFG with their SSA
names.
}
\details{
Generally, this function should only be called from \code{to_ssa()}.
}
