% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_ssa.R
\name{ssa_rename_ast}
\alias{ssa_rename_ast}
\title{Rename AST Variables with SSA Names}
\usage{
ssa_rename_ast(node, builder, paramNames = character())
}
\arguments{
\item{node}{(ASTNode) An abstract syntax tree.}

\item{builder}{(SSABuilder) A stateful object used by the renaming
algorithm.}
}
\description{
This function renames variables in an AST with their SSA names.
}
\details{
Generally, this function should only be called from \code{ssa_rename()}.
}
