% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_ssa.R
\name{ssa_set_numbers}
\alias{ssa_set_numbers}
\title{Set SSA Numbers on Symbols}
\usage{
ssa_set_numbers(node, cfg, dom_t, blocks, active, counter, ...)
}
\arguments{
\item{node}{(Function) The code object to number.}

\item{cfg}{(igraph) The control flow graph.}

\item{dom_t}{(igraph) The dominator tree for the CFG.}

\item{blocks}{(list) The blocks in the control flow graph.}

\item{active}{(integer) A vector of active SSA numbers, set by previous
blocks.}

\item{counter}{(Counter) A counter to generate unique SSA numbers.}

\item{...}{Additional arguments to methods.}
}
\description{
This function sets the SSA numbers on Symbols in code.
}
\details{
Generally, this function should only be called from \code{to_ssa()}.
}
