% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph.R
\name{successors}
\alias{successors}
\title{Vertex Successors}
\usage{
successors(g, id)
}
\arguments{
\item{g}{(igraph) The graph.}

\item{id}{An index into the graph, or a Block.}
}
\description{
This function gets the successors of a vertex in a graph.
}
\details{
This is a utility function to make working with igraph easier.
}
