% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_ast.R
\name{to_ast_}
\alias{to_ast_}
\title{Convert Quoted R Code to ASTNodes}
\usage{
to_ast_(expr, parent = NULL)
}
\arguments{
\item{expr}{(language) quoted R code to be converted.}
}
\description{
This function converts quoted R code to a tree of ASTNode objects. In other
words, this is the standard evaluation version of to_ast().
}

