% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_ast.R
\name{to_ast_callable}
\alias{to_ast_callable}
\title{Convert a Callable Object to an ASTNode}
\usage{
to_ast_callable(expr, is_primitive = FALSE)
}
\arguments{
\item{expr}{(language) Quoted R code to be converted.}
}
\description{
This function converts primitives, functions, and function definitions to
ASTNode objects. No distinction is made between functions and function
definitions since they have the same underlying code.
}
