% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_ast.R
\name{to_ast_function_def}
\alias{to_ast_function_def}
\title{Convert a function definition to an ASTNode.}
\usage{
to_ast_function_def(expr, parent = NULL)
}
\description{
Convert a function definition to an ASTNode.
}

