% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_ast.R
\name{to_ast_repeat}
\alias{to_ast_repeat}
\title{Convert a repeat to an ASTNode}
\usage{
to_ast_repeat(expr)
}
\arguments{
\item{expr}{(language) Quoted R code to be converted.}
}
\description{
Convert a repeat to an ASTNode
}
