% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_ast.R
\name{to_astq}
\alias{to_astq}
\title{Convert Unquoted R Code to ASTNodes}
\usage{
to_astq(expr, parent = NULL)
}
\arguments{
\item{expr}{unquoted R code to be converted.}
}
\description{
This function uses non-standard evaluation to convert unquoted R code to a
tree of ASTNode objects.
}

