% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_cfg.R
\name{to_basic_blocks}
\alias{to_basic_blocks}
\title{Build Basic Blocks from ASTNodes}
\usage{
to_basic_blocks(node, cfg = CFGraph$new())
}
\arguments{
\item{node}{(ASTNode) An ASTNode to build basic blocks from.}

\item{cfg}{(CFGraph) A CFG to insert the basic blocks into.}
}
\description{
This helper function does a depth-first traversal of an AST in order to
build basic blocks for a CFG.
}
\details{
Generally, this function should only be called from \code{to_cfg()}.
}
