% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_blocks.R
\name{to_blocks}
\alias{to_blocks}
\title{Convert an ASTNode Object to Basic Blocks}
\usage{
to_blocks(node, in_place = FALSE, ssa = TRUE, insert_return = TRUE)
}
\arguments{
\item{node}{A quoted R expression or an abstract syntax tree.}

\item{in_place}{(logical) Operate on the node in place? If \code{TRUE}, the
node will be modified.}

\item{ssa}{(logical) Also convert to static single assignment form?}

\item{insert_return}{(logical) Apply \code{insert_return()} to the node
before generating the CFG?}
}
\value{
A Function node with the control flow graph in its \code{$cfg}
field.
}
\description{
This function converts an ASTNode object or a quoted R expression into a
list of basic blocks. When the root of the AST is a Function object, the
CFG is built for its body.
}
\details{
Each basic block contains a sequence of non-branching program instructions.
Each basic block ends with an instruction that branches to one or more other
basic blocks.
}
