% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_cfg.R
\name{to_cfg}
\alias{to_cfg}
\title{Generate Control Flow Graph for a Function}
\usage{
to_cfg(fn)
}
\arguments{
\item{func}{(Function) a Function to generate the graph for}
}
\description{
This function generates the control flow graph for a Function object.
}

