% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_r.R
\name{to_r}
\alias{to_r}
\title{Convert ASTNode to R Code}
\usage{
to_r(node, ...)
}
\arguments{
\item{node}{(ASTNode) The tree to be converted.}

\item{keep_ssa}{(logical) Use SSA numbering?}

\item{use_phi}{(logical) Include phi-functions?}

\item{use_inner_blocks}{(logical) Include inner blocks?}
}
\description{
This function convert an abstract syntax tree to the corresponding R code.
}
