% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_ssa.R
\name{to_ssa}
\alias{to_ssa}
\title{Convert CFGraph to Static Single-Assignment Form}
\usage{
to_ssa(cfg, in_place = FALSE, renameParams = FALSE)
}
\arguments{
\item{cfg}{(CFGraph) A control flow graph.}

\item{in_place}{(logical) Don't copy CFG before conversion?}

\item{renameParams}{(logical) if TRUE, the parameter names are mapped to SSA form, e.g., x becomes x_1. Otherwise, the parameter names remain unaltered in the CFG.}
}
\value{
The control flow graph as a CFGraph object, with the code in each
block converted to SSA form.
}
\description{
This function converts code in a control flow graph (CFG) to static
single-assignment form.
}
