% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_ast.R
\name{toAST}
\alias{toAST}
\title{Convert Quoted R Code to ASTNodes}
\usage{
toAST(expr)
}
\arguments{
\item{expr}{(language) Quoted R code to be converted.}
}
\description{
This function converts quoted R code to a tree of ASTNode objects.
}
