% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_ast.R
\name{toASTCallable}
\alias{toASTCallable}
\title{Convert a Callable Object to an ASTNode}
\usage{
toASTCallable(expr, is_primitive = FALSE)
}
\arguments{
\item{expr}{(language) Quoted R code to be converted.}

\item{is_primitive}{(logical) Whether or not the expression is a primitive.}
}
\description{
This function converts primitives, functions, and function definitions to
ASTNode objects. No distinction is made between functions and function
definitions since they have the same underlying code.
}
