% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_ast.R
\name{toASTRepeat}
\alias{toASTRepeat}
\title{Convert a repeat to an ASTNode}
\usage{
toASTRepeat(expr)
}
\arguments{
\item{expr}{(language) Quoted R code to be converted.}
}
\description{
Convert a repeat to an ASTNode
}
