% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_cfg.R
\name{toCFGq}
\alias{toCFGq}
\title{Build Control Flow Graph from R Expression}
\usage{
toCFGq(expr, ...)
}
\arguments{
\item{expr}{An unquoted R expression.}

\item{...}{Additional arguments to \code{toCFG()}.}
}
\description{
This function builds the control flow graph (CFG) for an unquoted R
expression.
}
