% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_r.R
\name{toR}
\alias{toR}
\title{Convert ASTNodes to R Code}
\usage{
toR(node)
}
\arguments{
\item{node}{(ASTNode) The tree to be converted.}
}
\description{
This function converts a tree of ASTNodes to the corresponding R code.
}
