% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_ssa.R
\name{toSSA}
\alias{toSSA}
\title{Convert CFGraph to Static Single-Assignment Form}
\usage{
toSSA(node, inPlace = FALSE)
}
\arguments{
\item{node}{(Function) A function with CFG attached.}

\item{inPlace}{(logical) Don't copy CFG before conversion?}
}
\value{
The function with its control flow graph converted to static
single-assignment form.
}
\description{
This function converts code in a control flow graph (CFG) to static
single-assignment form.
}
