% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{toString.ASTNode}
\alias{toString.ASTNode}
\alias{toString.Label}
\title{Convert an ASTNode to a Character String}
\usage{
\method{toString}{ASTNode}(x, ...)

\method{toString}{Label}(x, ..., block_prefix = "\%")
}
\arguments{
\item{x}{(ASTNode) The object to be converted.}

\item{...}{Optional arguments passed to or from methods.}

\item{block_prefix}{(character) Prefix to prepend to basic block IDs.}
}
\value{
A character vector of length 1.
}
\description{
These methods produce a single string that describes the code an ASTNode
represents.
}
\examples{
ast = quote_ast(if (x > 3) 42 else "hello")
toString(ast)
}
