% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{transform_lines}
\alias{transform_lines}
\title{Transform Elements in a \code{Container} Object}
\usage{
transform_lines(node, transform, ...)
}
\arguments{
\item{node}{(ASTNode) The node to transform.}

\item{transform}{(function) The transformation function to apply to each
line. See the details section.}
}
\value{
The \code{node} argument after transformation.
}
\description{
This function transforms each element in a \code{Container} object by applying
the given \code{transform} function.
}
\details{
This is an in-place transformation!

This function is intended for transforming lines in \code{Block} objects, but
also works with other \code{Container} objects.

This function can transform a single element in a container into multiple
elements. If the \code{transform} function returns a list, the original element
is replaced with one element for each element in the list.

If the \code{transform} function returns an ASTNode, the original element is
replaced with that node.

The index of the original element is passed to the \code{transform} function as
\code{index}. If the top-level \code{node} argument was a \code{BlockList} object, the
block index is also passed to the \code{transform} function as \code{block_index}.
}
\seealso{
\code{\link[=replace_nodes]{replace_nodes()}}
}
