% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_blocks.R
\name{as.data.frame.FunctionBlocks}
\alias{as.data.frame.FunctionBlocks}
\alias{as_data_frame}
\alias{as_data_frame.FunctionBlocks}
\title{Convert Basic Blocks to a Code Data Frame}
\usage{
\method{as.data.frame}{FunctionBlocks}(x, ...)

as_data_frame(x, ...)

\method{as_data_frame}{FunctionBlocks}(x, ...)
}
\arguments{
\item{x}{(FunctionBlocks) The basic blocks to convert.}

\item{...}{Additional arguments to be passed to or from methods.}
}
\description{
This function converts basic blocks to a data frame where each row is one
``line'' of code.
}
