% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_blocks.R
\name{as.data.frame.BlocksList}
\alias{as.data.frame.BlocksList}
\alias{as_data_frame}
\alias{as_data_frame.BlocksList}
\title{Convert Basic Blocks to a Code Data Frame}
\usage{
\method{as.data.frame}{BlocksList}(x, ...)

as_data_frame(x, ...)

\method{as_data_frame}{BlocksList}(x, ...)
}
\arguments{
\item{x}{(BlocksList) The basic blocks to convert.}

\item{...}{Additional arguments to be passed to or from methods.}
}
\description{
This function converts basic blocks to a data frame where each row is one
``line'' of code.
}
