% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_nodes.R
\name{find_nodes}
\alias{find_nodes}
\title{Find Nodes in an ASTNode Object}
\usage{
find_nodes(node, check, ..., recursive = TRUE, .address = integer(0))
}
\arguments{
\item{node}{(ASTNode) The code object to search.}

\item{check}{(function) The check function, which should take the ASTNode
object to check as its first argument and return a scalar logical.}

\item{...}{Additional arguments to the check function.}

\item{recursive}{(logical) Search recursively in matching nodes?}
}
\value{
A list of addresses for matching nodes.
}
\description{
This function returns a list of addresses for all children of an ASTNode
object that satisfy the given check function.
}
\examples{
ast = quote_ast(
  function(x, y) {
    x = x + y
    function() x
  })

find_nodes(ast, is, "Function")

find_nodes(ast, function(node) is(node, "Symbol") && node$name == "x")
}
