% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_nodes.R
\name{get_index}
\alias{get_index}
\title{Get Index of RStatic Object}
\usage{
get_index(node, index = integer(0))
}
\arguments{
\item{node}{(ASTNode) The RStatic object to traverse.}

\item{index}{(integer) Initial index. The value of this parameter is
appended to the end of the computed index.}
}
\value{
An integer vector which can be used as an index in the \code{child}
function.
}
\description{
This function gets the index of the given RStatic object relative to its
ancestor nodes.
}
\examples{
ast = quote_ast(x <- y + z)
node = child(ast, c(2, 2, 1))

get_index(node)
}
