% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph.R
\name{postorder}
\alias{postorder}
\title{Postorder of a Graph}
\usage{
postorder(g, root)
}
\arguments{
\item{g}{(igraph) The graph.}

\item{root}{(integer) A vertex where the traversal should start.}
}
\value{
(integer) A postorder of the graph.
}
\description{
Given a graph, this function returns the order the vertices are exited in a
depth-first traversal.
}
\details{
A postorder is not always unique. For an acyclic graph, reversed postorder
is equivalent to a topological sort.
}
