% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfg_analysis.R
\name{postorder}
\alias{postorder}
\title{Postorder Traversal of a Graph}
\usage{
postorder(cfg, from = cfg$entry)
}
\arguments{
\item{cfg}{(CFGraph) The graph on which to compute a traversal.}

\item{from}{(integer) A node index where the traversal should start.}
}
\value{
(integer) The node indexes in the order they'd be visited.
}
\description{
Given a graph, this function finds the order nodes are visited by a
postorder traversal.
}
