% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_nodes.R
\name{replace_nodes}
\alias{replace_nodes}
\title{Replace Nodes in an ASTNode Object}
\usage{
replace_nodes(node, fn, ...)
}
\arguments{
\item{node}{(ASTNode) The code object to modify.}

\item{fn}{(function) The function to call on each node.}
}
\description{
This function replaces each node of an ASTNode object (including the object
itself) with the result of calling the given function on the node.
}
\details{
Nodes are replaced from the bottom up, so the given node is replaced last.
}
\examples{
rename_symbols = function(node, name, newname)
{ # Rename symbols to something else.
  if (is(node, "Symbol") && node$basename == name)
    node$basename = newname

  node
}

ast = quote_ast(ans <- sum(x, y / x, z))

replace_nodes(ast, rename_symbols, "x", "newx")
}
