% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_ssa.R
\name{ssa_rename}
\alias{ssa_rename}
\title{Rename CFG Variables with SSA Names}
\usage{
ssa_rename(block, cfg, dom_t, ns = NameStack$new())
}
\arguments{
\item{block}{(integer) Identifier of a basic block in the CFG.}

\item{cfg}{(CFGraph) A control-flow graph.}

\item{dom_t}{(integer) The dominator tree for the CFG.}

\item{ns}{(NameStack) A stateful object used by the renaming algorithm.}
}
\description{
This function renames variables in the basic blocks of a CFG with their SSA
names.
}
\details{
Generally, this function should only be called from \code{to_ssa()}.
}

