% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_ast.R
\name{to_ast}
\alias{to_ast}
\title{Convert R Code to ASTNodes}
\usage{
to_ast(expr)
}
\arguments{
\item{expr}{(language) The R code to be converted.}
}
\description{
This function converts quoted R code to the corresponding tree of ASTNodes.
}

