% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_ast.R
\name{to_ast_function_def}
\alias{to_ast_function_def}
\title{Convert a Function Definition to an ASTNode}
\usage{
to_ast_function_def(expr)
}
\arguments{
\item{expr}{(language) Quoted R code to be converted.}
}
\description{
Convert a Function Definition to an ASTNode
}
