% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_r.R
\name{to_r}
\alias{to_r}
\title{Convert Intermediate Representations to R Code}
\usage{
to_r(node, ...)
}
\arguments{
\item{node}{The intermediate representation to convert.}

\item{...}{Optional arguments to and from methods.}
}
\description{
This function converts rstatic intermediate representations back to R code.
}
