% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_ssa.R
\name{to_ssa}
\alias{to_ssa}
\title{Convert Blocks to Static Single-Assignment Form}
\usage{
to_ssa(node, cfg = compute_cfg(node), dom_t = dominator_tree(cfg, 2),
  in_place = FALSE, ...)
}
\arguments{
\item{node}{A Function or BlocksList object.}

\item{cfg}{(igraph) The control flow graph for the code.}

\item{dom_t}{(igraph) The dominator tree for the code.}

\item{in_place}{(logical) Modify the node in-place? If \code{FALSE}, the
node is copied.}

\item{...}{Additional arguments to methods.}
}
\value{
NULL, because this function modifies the blocks in place.
}
\description{
This function converts code blocks to static single-assignment form.
}
\details{
This function modifies the blocks in place.
}
