% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_ssa.R
\name{to_ssa}
\alias{to_ssa}
\title{Convert CFGraph to Static Single-Assignment Form}
\usage{
to_ssa(node, in_place = FALSE)
}
\arguments{
\item{node}{(Function) A function with CFG attached.}

\item{in_place}{(logical) Don't copy CFG before conversion?}
}
\value{
The function with its control flow graph converted to static
single-assignment form.
}
\description{
This function converts code in a control flow graph (CFG) to static
single-assignment form.
}
